Reverted non working merge commit
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/branches/SGppAutoTune@5454 4eea3252-f0fb-4393-894d-40516dce545b

+++ datadriven/src/sgpp/datadriven/application/BatchLearner.cpp
@@ -14,7 +14,7 @@
 #include <deque>
 #include <algorithm>
 
-#include <sgpp/datadriven/application/BatchLearner.hpp>
+#include "BatchLearner.hpp"
 #include <sgpp/base/datatypes/DataVector.hpp>
 #include <sgpp/base/grid/Grid.hpp>
 #include <sgpp/base/grid/GridStorage.hpp>
@@ -119,8 +119,7 @@ namespace SGPP {
 
       //iterate over all found lines
       for (size_t i = 0; i < data.size() - 1; i++) {
-        size_t classesHere = static_cast<size_t>(
-                               count(data[i].begin(), data[i].end(), ','));
+        unsigned long int classesHere = count(data[i].begin(), data[i].end(), ',');
 
         // the first data entry ever found determines the number of dimensions of the data
         if (dimensions == 0) {
